+++ engine/src/ui/GridLayout.hpp
@@ -7,8 +7,8 @@ namespace Birdy3d::ui {
     class StaticGridLayout : public Layout {
     public:
         StaticGridLayout(int gap = 0);
-        void arrange(const std::list<std::shared_ptr<Widget>>& children, glm::vec2 pos, glm::vec2 size) const override;
-        glm::vec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children) const override;
+        void arrange(const std::list<std::shared_ptr<Widget>>& children, glm::ivec2 pos, glm::ivec2 size) const override;
+        glm::ivec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children) const override;
 
     private:
         int m_gap;
@@ -17,9 +17,9 @@ namespace Birdy3d::ui {
     class DynamicGridLayout : public Layout {
     public:
         DynamicGridLayout(int gap = 0);
-        void arrange(const std::list<std::shared_ptr<Widget>>& children, glm::vec2 pos, glm::vec2 size) const override;
-        glm::vec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children) const override;
-        glm::vec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children, float suggested_size, Direction) const override;
+        void arrange(const std::list<std::shared_ptr<Widget>>& children, glm::ivec2 pos, glm::ivec2 size) const override;
+        glm::ivec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children) const override;
+        glm::ivec2 minimal_size(const std::list<std::shared_ptr<Widget>>& children, float suggested_size, Direction) const override;
         float minimal_size(const std::list<std::shared_ptr<Widget>>& children, Direction) const override;
 
     private:
