+++ engine/src/ui/Layout.cpp
@@ -4,7 +4,7 @@
 
 namespace Birdy3d::ui {
 
-    glm::vec2 Layout::minimal_size(const std::list<std::shared_ptr<Widget>>& children, float, Layout::Direction) const {
+    glm::ivec2 Layout::minimal_size(const std::list<std::shared_ptr<Widget>>& children, float, Layout::Direction) const {
         return minimal_size(children);
     };
 
@@ -12,30 +12,30 @@ namespace Birdy3d::ui {
         return direction == Layout::Direction::HORIZONTAL ? minimal_size(children).x : minimal_size(children).y;
     }
 
-    void MaxLayout::arrange(const std::list<std::shared_ptr<Widget>>& children, glm::vec2 pos, glm::vec2 size) const {
+    void MaxLayout::arrange(const std::list<std::shared_ptr<Widget>>& children, glm::ivec2 pos, glm::ivec2 size) const {
         for (auto it = children.rbegin(); it != children.rend(); it++) {
             (*it)->arrange(pos, size);
         }
     }
 
-    glm::vec2 MaxLayout::minimal_size(const std::list<std::shared_ptr<Widget>>& children) const {
+    glm::ivec2 MaxLayout::minimal_size(const std::list<std::shared_ptr<Widget>>& children) const {
         if (children.empty())
-            return glm::vec2(0);
+            return glm::ivec2(0);
         return children.front()->minimal_size();
     }
 
-    void AbsoluteLayout::arrange(const std::list<std::shared_ptr<Widget>>& children, glm::vec2 pos, glm::vec2 size) const {
+    void AbsoluteLayout::arrange(const std::list<std::shared_ptr<Widget>>& children, glm::ivec2 pos, glm::ivec2 size) const {
         for (const auto& child : children) {
-            glm::vec2 child_size = child->preferred_size(size);
-            glm::vec2 child_pos = pos + child->preferred_position(size, child_size);
+            glm::ivec2 child_size = child->preferred_size(size);
+            glm::ivec2 child_pos = pos + child->preferred_position(size, child_size);
             if (child_size.x == 0 || child_size.y == 0)
                 child_size = size;
             child->arrange(child_pos, child_size);
         }
     }
 
-    glm::vec2 AbsoluteLayout::minimal_size(const std::list<std::shared_ptr<Widget>>& children [[maybe_unused]]) const {
-        return glm::vec2(0);
+    glm::ivec2 AbsoluteLayout::minimal_size(const std::list<std::shared_ptr<Widget>>& children [[maybe_unused]]) const {
+        return glm::ivec2(0);
     }
 
 }
